API additions (classes + methods) to make it easier to work with Hap data in AVFoundation.
AVPlayerItemHapDXTOutput can now output RGBA data.  Added method for synchronously fetching and decoding hap frames more suitable for transcoding/fetching thumbnails/etc.  Added a method for decoding arbitrary CMSampleBufferRefs containing Hap data obtained elsewhere (from AVAssetReader, for example).  Added AVAssetReaderHapTrackOutput class to simplify fetching RGB-based sample buffers from hap movies loaded as AVAssets.  Updated the sample app to demonstrate CPU-based decoding of hap frames to RGBA NSImages.

+++ HapInAVF Test App/HapPixelBufferTexture.m
@@ -116,7 +116,7 @@
 	// Ignore the value for CVPixelBufferGetBytesPerRow()
 	
 	size_t bytesPerRow = (roundedWidth * bitsPerPixel) / 8;
-	GLsizei newDataLength = bytesPerRow * roundedHeight; // usually not the full length of the buffer
+	GLsizei newDataLength = (int)(bytesPerRow * roundedHeight); // usually not the full length of the buffer
 	
 	size_t actualBufferSize = [decodedFrame dxtDataSize];
